Iteration 1 Perlin noise driven position change
In order to better display the animation effect, this step changes the way the canvas is centered, so the drawing logic of some shapes is changed, and the results produced are adjusted to achieve the original effect. At the same time, some elements in the picture (such as the semicircle) are deleted.